requirements.txt
-i https://pypi.python.org/simple
asgiref==3.5.2 ; python_version >= '3.7'
astroid==2.12.12 ; python_full_version >= '3.7.2'
asttokens==2.0.8
attrs==22.1.0 ; python_version >= '3.5'
autoflake==1.7.6
autopep8==1.7.0
backcall==0.2.0
black==22.10.0
certifi==2022.9.24 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
decorator==5.1.1 ; python_version >= '3.5'
dill==0.3.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
django==4.1.2
django-debug-toolbar==3.7.0
django-debugtools==2.0
django-extensions==3.2.1
django-pudb3==0.2.0
docformatter==1.5.0 ; python_version >= '3.6' and python_version < '4.0'
executing==1.1.1
flake8==5.0.4
greenlet==1.1.3.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.4 ; python_version >= '3.5'
iniconfig==1.1.1
ipython==8.5.0
isort==5.10.1 ; python_version < '4.0' and python_full_version >= '3.6.1'
jedi==0.18.1 ; python_version >= '3.6'
lazy-object-proxy==1.7.1 ; python_version >= '3.6'
matplotlib-inline==0.1.6 ; python_version >= '3.5'
mccabe==0.7.0 ; python_version >= '3.6'
msgpack==1.0.4
mypy-extensions==0.4.3
neovim==0.3.1
packaging==21.3 ; python_version >= '3.6'
parso==0.8.3 ; python_version >= '3.6'
pathspec==0.10.1 ; python_version >= '3.7'
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
platformdirs==2.5.2 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
prompt-toolkit==3.0.31 ; python_full_version >= '3.6.2'
ptyprocess==0.7.0
pudb==2022.1.2
pure-eval==0.2.2
py==1.11.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycodestyle==2.9.1 ; python_version >= '3.6'
pyflakes==2.5.0 ; python_version >= '3.6'
pyformat==0.7
pygments==2.13.0 ; python_version >= '3.6'
pylint==2.15.4
pynvim==0.4.3
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
pytest==7.1.3
pytest-django==4.5.2
requests==2.28.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
stack-data==0.5.1
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1 ; python_version < '3.11'
tomlkit==0.11.5 ; python_version >= '3.6' and python_version < '4.0'
traitlets==5.5.0 ; python_version >= '3.7'
unify==0.5
untokenize==0.1.1
urllib3==1.26.12 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
urwid==2.1.2
urwid-readline==0.13
wcwidth==0.2.5
wrapt==1.14.1 ; python_version < '3.11'
yapf==0.32.0
django-filter==22.1
djangorestframework==3.14.0
psycopg2-binary==2.9.4
pytz==2022.5
